Export sentinel package errors
Before this commit, the only way to explicitly check which error occured
was to test against the actual string error message.
This is a anti-pattern and comes with a high risk of breaking when the
value gets changed in the library.

To allow to check against the errors, this commit exports the types so
they can be tested using the `errors.Is(error, error)` function from
Go's `errors` package.

A suggestion to implement this change was already requested in
https://github.com/hako/branca/issues/5#issuecomment-416915978,
but at the time of this commit a PR was never submitted.

This commit also includes 4 new tests to ensure the exported error types
can be checked using `errors.Is(error, error)`.
The execution of the tests is way slower because the new
`TestExpiredTokenError` test makes use of the
`time.Sleep(time.Duration)` function with a value of 3 seconds in order
to ensure the TTL of the token is expired, but unfortunately there is no
other way to test this scenario.