Implement ErrExpiredToken as error type with additional metadata
Implemented the `ErrExpiredToken` error as new error type and added more
details about the occurred error. This allows the consumer to get the
expiration time of the token as `time.Time` value. The error message
itself was extended by adding the relative time in `time.Duration`
format, e.g. "token is expired by 1m2s".
The `ErrExpiredToken.Time` field can be used by consumers to do more
calucations or convert and format the value to a specific timezone.