Remove inappropriate error returned from implementation dependency
The `github.com/eknkc/basex` package is internal library dependency
and therefore an implementation detail.
The `DecodeToString(string) (string, error)` method could error when the
call to the `basex.NewEncoding(string)` method returned an error that
indicates an invalid alphabet.

Before an `ErrInvalidToken` error was returned which might confuse the
consumer since this could also occur even when the token is valid.
This commit changes the returned error to only provide the error message
of the original error that might indicate a critical error not related
to the library. THe consumer can now decide how to handle this error.